chore: Add fillable properties to Apartment and Owner models
This commit adds the `fillable` properties to the `Apartment` and `Owner` models, allowing mass assignment for the specified attributes. This change improves the flexibility and ease of use when creating new instances of these models.

Note: This commit message follows the established convention of using the `chore` type for non-functional changes that don't affect the behavior of the code.